Moved onboarding into modules and added translation
