+++ public/locales/cn/auth.json
@@ -74,11 +74,35 @@
         "invalid-email": "请输入有效邮箱"
       }
     },
-    "onboarding":{
-      "pageTitle": "上岸了哈哈哈！",
-    },
+   
     "error": "这个邮箱不存在呀",
     "success": "我们收到你的重置密码请求啦，请查收邮箱",
     "button": "发送邮件"
+  },
+  "onboarding":{
+    "pageTitle": "上岸了哈哈哈！",
+    "introduction":"宝，自我介绍一下呗？",
+    "service":"咱这蹭蹭不花钱！",
+    "askName":"宝贝名字留一个呀？",
+    "category":"宝贝喜欢啥样的?",
+    "continue":"继续不要停~",
+    "error": {
+      "required": "小宝贝名字别忘了呀",
+      "invalid-name": "宝贝不舍得告诉我真名吗？"
+    },
+    "tags":{
+      "Business":"🏢商务黑丝",
+      "Creative":"🎨创造力爆棚",
+       "Education":"📚教育的小皮鞭",
+       "Entertainment":"🎶开开心心娱乐一下",
+       "Fasion":"👗靓仔靓女",
+       "Food":"🍕吃货",
+       "Government":"⚖️妄议国事",
+       "Health":"🍎保温杯里泡枸杞",
+       "Non-profit":"💗视金钱为闰土",
+       "Other":"💗还有谁？",
+       "Tech":"🖥第一生产力",
+       "Travel":"✈️诗和远方"
+    }
   }
 }
