+++ public/locales/en/auth.json
@@ -73,11 +73,34 @@
         "invalid-email": "Please enter a valid email address"
       }
     },
-    "onboarding":{
-      "pageTitle": "onboarding"
-    },
     "error": "Oops! We can't find that email.",
     "success": "We have received your request, please check your email!",
     "button": "Send Email"
+  },
+  "onboarding":{
+    "pageTitle": "Welcome Aboard!",
+    "introduction":"Tell Us About Yourself?",
+    "service":"Free forever. No payment needed!",
+    "askName":"Tell us your name?",
+    "category":"Select one category that best descrbes your A-Comosus:",
+    "continue":"Continue",
+    "error": {
+      "required": "Name is required",
+      "invalid-email": "Please enter a valid email address"
+    },
+    "tags":{
+      "Business":"🏢Business",
+      "Creative":"🎨Creative",
+      "Education":"📚Education",
+      "Entertainment":"🎶Entertainment",
+      "Fasion":"👗Fasion & Beauty",
+      "Food":"🍕Food & Beverage",
+      "Government":"⚖️Government & Politics",
+       "Health":"🍎Health & Wellness",
+       "Non-profit":"💗Non-profit",
+       "Other":"💗Other",
+       "Tech":"🖥Tech",
+       "Travel":"✈️Travel & Tourism"
+    }
   }
 }
