+++ src/modules/auth/index.ts
@@ -1,3 +1,4 @@
 export { default as LoginForm } from './LoginForm/LoginForm';
 export { default as ForgetPasswordForm } from './ForgetPasswordForm/ForgetPasswordFrom';
 export { default as SignUpForm } from './SignUpForm/SignUpForm';
+export { default as OnboardingPage } from './Onboarding/OnboardingPage';
